clean up parser registry
